OML😭😭😭 [GRAPH]/[DESIGNER] BUTTONS FINALLY IN!!!
i don wanna talk about it

jk i do

I ran into way more hoops than I'd like to admit lmao

I wanted to replicate the look and feel of UMG's graph and designer toolbar buttons 1:1. This means 1) they don't collapse in the overflow menu and 2) they're justified in the righthand corner.

However, by default, toolbar buttons are forced into left justification and don't take priority over collapse behavior. This *would* be simple to fix using FExtender and/or subclassing!!

But I'm designing the plugin to inject into ANY actor class. So I cannot override this default behavior using normal means. I *have* to work around blueprints' default application mode and toolbar behavior. I'm working around the public API, too, which limits me on using internal functions to inject the buttons.

On top of that, OnAssetEditorOpened() is too late to even inject in FExtender if I wanted to. The UI is locked in by then so any changes silently fail. OnEditorOpeningPreWidgets() IS early enough, but the blueprint isn't fully initialized so FExtender runs into un-initialized ptr crashing territory. And that's all while ignoring how I'd have to subclass, which no🙅!

So what's the solution🫣?!?! To use the widget reflector on the toolbar and find exactly where it's UI lives inside blueprints🫠! This is under 'SStandaloneAssetEditorToolkitHost', inside of a 'SHorizontalBox'. So, I just need to inject the buttons there!

Boom! Just search for that unique string and walk down the tree a tiny bit, injecting the buttons during OnEditorPreWidgets() before UI is locked in!

Adding it into the horizontal box is perfect because it lives *beside* the toolbar, not collapsing in the overflow menu! AND it gives me the freedom to justify the buttons on the right side!! win-win!

ANNND IT LOOKS AND FUNCTIONS *EXACTLY* THE SAME AS UMG SO THAT'S KINDA SICK DUDE!!!

I'll move the logic outta the subclass later i just want it pushed now lol

TL;DR:  that was a pain !