⛔ Fixed inability to re-open assets ⛔ (yeesh)
It was a fatal bug that I ignored for a little while since I was focused on getting the actual editor to function

Now that the editor works this was the first thing I knew I needed to take care of 🫣🫨

In hindsight it was a little silly, it was just dangling pointers ahah! I kinda assumed the engine would've handled the fatal stuff for me. I still feel like it should've🤨?

Maybe it's because I'm trying to blend modern systems into this super old architecture that in-itself has had a ton of modernized patches🧩. Kinda like how I'm replacing AssetTypeActions with modern Definitions.

Backwards compatibility can be a pain to work around!! I actually had noticed that FBlueprintEditor uses a function from the DMX Editor which is misspelled: 'InitalizeExtenders()'🔥🔥🔥. It made me giggle 'cause this decade-old function probably had to be grandfathered in because of backwards-compatibility😝

But, regardless, it's better to track lifetime myself from now on. bad on me oops