Stack | stackplugin.com !
This plugin, PanelDesigner, has turned into : Stack!

I bought the domain stackplugin.com to lock things in. (idk how the hell that domain wasn't taken 'cause that's a kickass URL ah!!🤩)

Some reasons that stirred this pivot~

I was going to try and re-use Niagara's mvvm architecture, but I don't actually have access to those modules in binary engine builds. That's a hard line for me since I want to make this plugin accessible for all devs (for using source engine and binary). Regardless, it'd also be a bit bloat-y having to pull in such a complex/large editor just to re-integrate a sliver of it.

So I decided to pivot this plugin into becoming an ecosystem! My framework for building my editor is now spun into its own module, StackFramework, which I'm now designing as a generalized API that devs can also use for making their own stack-related things!

Right now modules are split into:
- StackFramework (API)
- StackDesigner (Panel designer, integrated into *other* editors)
- StackEditor (Standalone editor for creating stack UI asset types)
- StackNodes (Runtime module for K2 stack-nodes)

Ik It's a bold pivot, and it's gonna be a lot of work, but I believe in its value :')