Placeable Stack node w/ dummy entry list 🥳📃🎊 !!
Every class in Phase I of Milestone I is hooked up with one another- 🥳Placing a stack node shows 4 dummy entries shown in a list! ah!!!

Right now it's mostly just the slate classes hooked up with a few vital ViewModel classes holding dummy data. This is the flow of how the classes are working with one another:

SStackNode
└── SStackView
    └── SStackList
        └── SStackEntry
            └── SStackRow
                └── SStackItem